requirements.txt
Pillow>=9.0.0 
psutil>=5.9.0
matplotlib>=3.5.0 